# api-spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Optimus AI Platform API",
    "description": "Full REST API specification for the Optimus AI platform covering authentication, users, chats, models, files, knowledge, prompts, tools, groups, functions, skills, memories, notes, channels, automations, calendar, folders, configs, and third-party integrations.",
    "version": "1.0.0",
    "contact": {
      "name": "Optimus AI Support",
      "url": "https://github.com/optimus-ai"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "security": [
    { "BearerAuth": [] },
    { "ApiKeyAuth": [] }
  ],
...
  "paths": {
    "/api/config": {
      "get": {
        "tags": ["Config"],
        "summary": "Get global configuration",
        "operationId": "getGlobalConfig",
        "responses": {
          "200": { "description": "Configuration retrieved", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfigResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/models": {
      "get": {
        "tags": ["Models"],
        "summary": "List available models",
        "operationId": "listModels",
        "responses": {
          "200": { "description": "Models listed", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Model" } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/chat/completions": {
      "post": {
        "tags": ["Chat"],
        "summary": "Create chat completion",
        "operationId": "createChatCompletion",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatCompletionRequest" } } }
        },
        "responses": {
          "200": { "description": "Completion created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatCompletionResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
    "/api/v1/models/model": {
      "get": {
        "tags": ["ModelAdmin"],
        "summary": "Get model details",
        "operationId": "getModel",
        "parameters": [
          { "name": "id", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Model", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Model" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/openai/config/update": {
      "post": {
        "tags": ["OpenAI"],
        "summary": "Update OpenAI config",
        "operationId": "updateOpenAIConfig",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OpenAIConfig" } } }
        },
        "responses": {
          "200": { "description": "Config updated" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token. Obtained via /api/v1/auths/signin or /api/v1/auths/signup."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "API key with 'sk-' prefix. Passed as 'Bearer sk-...'"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized – missing or invalid credentials",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string", "description": "Error message" }
        },
        "required": ["detail"]
      },
      "ConfigResponse": {
        "type": "object",
        "properties": {
          "version": { "type": "string" },
          "defaultModel": { "type": "string" },
          "features": { "type": "object", "additionalProperties": { "type": "boolean" } }
        }
      },
...
      "Model": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "object": { "type": "string", "default": "model" },
          "created": { "type": "integer" },
          "owned_by": { "type": "string" },
          "base_model_id": { "type": ["string", "null"] },
          "meta": { "type": "object", "additionalProperties": {} },
          "tags": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "ChatCompletionRequest": {
        "type": "object",
        "properties": {
          "model": { "type": "string" },
          "messages": { "type": "array", "items": { "$ref": "#/components/schemas/Message" } },
          "stream": { "type": "boolean", "default": false },
          "temperature": { "type": "number", "minimum": 0, "maximum": 2 },
          "top_p": { "type": "number", "minimum": 0, "maximum": 1 },
          "max_tokens": { "type": "integer" },
          "frequency_penalty": { "type": "number" },
          "presence_penalty": { "type": "number" },
          "tools": { "type": "array", "items": { "$ref": "#/components/schemas/ToolDefinition" } },
          "tool_choice": { "oneOf": [{ "type": "string" }, { "type": "object" }] }
        },
        "required": ["model", "messages"]
      },
      "ChatCompletionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "object": { "type": "string", "default": "chat.completion" },
          "created": { "type": "integer" },
          "model": { "type": "string" },
          "choices": { "type": "array", "items": { "$ref": "#/components/schemas/Choice" } },
          "usage": { "$ref": "#/components/schemas/Usage" }
        }
      },
      "Choice": {
        "type": "object",
        "properties": {
          "index": { "type": "integer" },
          "message": { "$ref": "#/components/schemas/Message" },
          "finish_reason": { "type": "string", "enum": ["stop", "length", "tool_calls"] }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "role": { "type": "string", "enum": ["system", "user", "assistant", "tool"] },
          "content": { "type": ["string", "null"] },
          "name": { "type": "string" },
          "tool_calls": { "type": "array", "items": { "$ref": "#/components/schemas/ToolCall" } },
          "tool_call_id": { "type": "string" }
        },
        "required": ["role"]
      },
      "ToolCall": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string", "default": "function" },
          "function": { "$ref": "#/components/schemas/FunctionCall" }
        }
      },
      "FunctionCall": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "arguments": { "type": "string" }
        }
      },
      "ToolDefinition": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "default": "function" },
          "function": { "$ref": "#/components/schemas/FunctionDefinition" }
        }
      },
      "FunctionDefinition": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "parameters": { "type": "object" }
        },
        "required": ["name"]
      },
      "Usage": {
        "type": "object",
        "properties": {
          "prompt_tokens": { "type": "integer" },
          "completion_tokens": { "type": "integer" },
          "total_tokens": { "type": "integer" }
        }
      },
...
      "OpenAIConfig": {
        "type": "object",
        "properties": {
          "apiKey": { "type": "string", "description": "OpenAI API key" },
          "baseUrl": { "type": "string", "format": "uri" }
        }
      },
...
    }
  }
}